Obtained a result!
Used the zoom feature in matplotlib to find the approximate root of the characteristic equation, then used a NR convergence to calculate it - not sure if it corresponds perfectly, real part doesnt quite but that is probably understandable because the graph only showed the real values, not the complex parts. However, complex part of our result is incredibly small, suggesting its just an artifact. Simulation does now show excellent mode localisation, however.